+++ src/graph/ontology_manager.cpp
@@ -214,7 +214,7 @@ static std::vector<YamlEntry> parseYamlSection(const std::vector<std::string> &l
         }
     };
 
-    while (i < lines.size()) {
+    while (static_cast<size_t>(i) < lines.size()) {
         const std::string &raw = lines[i];
         // count leading spaces
         int indent = 0;
@@ -565,7 +565,7 @@ bool OntologyManager::parseJson(const std::string &text) {
         return false;
     }
 
-    while (pos < text.size()) {
+    while (static_cast<size_t>(pos) < text.size()) {
         skipWs(text, pos);
         if (pos < text.size() && text[pos] == '}') {
             break;
@@ -657,7 +657,7 @@ bool OntologyManager::parseYaml(const std::string &text) {
     }
 
     std::size_t i = 0;
-    while (i < lines.size()) {
+    while (static_cast<size_t>(i) < lines.size()) {
         std::string raw = trimYaml(lines[i]);
         if (raw.empty() || raw[0] == '#') {
             ++i;
@@ -675,7 +675,7 @@ bool OntologyManager::parseYaml(const std::string &text) {
         // Determine indent of entries in this section
         int entry_indent = -1;
         std::size_t look = i;
-        while (look < lines.size()) {
+        while (static_cast<size_t>(look) < lines.size()) {
             std::string l = lines[look];
             int ind       = 0;
             while (ind < static_cast<int>(l.size()) && l[ind] == ' ') {
