+++ src/importers/mysql_importer.cpp
@@ -940,7 +940,7 @@ bool MySQLImporter::parseCreateTable(const std::string& sql, TableSchema& schema
         std::string col_type = {};
         size_t k = type_start;
         int tdep = 0;
-        while (k < col_def.size()) {
+        while (static_cast<size_t>(k) < col_def.size()) {
             char c = col_def[k];
             if (c == '(') { ++tdep; col_type += c; }
             else if (c == ')') {
@@ -1041,7 +1041,7 @@ bool MySQLImporter::parseInsert(const std::string& sql, const ImportOptions& opt
     // Walk the payload extracting one parenthesised tuple at a time.
     size_t pos = 0;
     size_t rows_imported = 0;
-    while (pos < values_payload.size()) {
+    while (static_cast<size_t>(pos) < values_payload.size()) {
         // Skip whitespace and commas between tuples
         while (pos < values_payload.size() &&
                (values_payload[pos] == ' ' || values_payload[pos] == '\t' ||
@@ -1049,7 +1049,7 @@ bool MySQLImporter::parseInsert(const std::string& sql, const ImportOptions& opt
                 values_payload[pos] == '\n')) {
             ++pos;
         }
-        if (pos >= values_payload.size()) {
+        if (pos >= static_cast<int>(values_payload.size())) {
           break;
         }
         if (values_payload[pos] != '(') {
@@ -1589,7 +1589,7 @@ std::string MySQLImporter::stripMySQLComments(const std::string& sql) {
     std::string result = {};
     result.reserve(sql.size());
     size_t i = 0;
-    while (i < sql.size()) {
+    while (static_cast<size_t>(i) < sql.size()) {
         if (i + 1 < sql.size() && sql[i] == '/' && sql[i + 1] == '*') {
             // Skip until closing */
             i += 2;
